move to modeling utils, integrate into existing source code
