+++ tests/test_big_modeling.py
@@ -24,7 +24,6 @@ import torch.nn as nn
 from transformers import AutoModelForCausalLM, AutoTokenizer
 
 from accelerate.big_modeling import (
-    align_module,
     cpu_offload,
     cpu_offload_with_hook,
     disk_offload,
@@ -1049,50 +1048,3 @@ class BigModelingTester(unittest.TestCase):
 
         assert model.h[0].self_attention.query_key_value.weight.dtype == torch.uint8
         assert model.h[0].self_attention.query_key_value.weight.device.index == 0
-
-    def test_align_module_simple(self):
-        model = ModelForTest()
-        execution_device = torch.device(torch_device)
-        model_device = torch.device("cpu")
-
-        # test default execution device
-        with align_module(model.batchnorm):
-            assert model.linear1.weight.device == model_device
-            assert model.batchnorm.weight.device == model_device
-            assert model.linear2.weight.device == model_device
-        assert model.linear1.weight.device == model_device
-        assert model.batchnorm.weight.device == model_device
-        assert model.linear2.weight.device == model_device
-
-        # test with explicit execution device
-        with align_module(model.batchnorm, execution_device=execution_device):
-            assert model.linear1.weight.device == model_device
-            assert model.batchnorm.weight.device == execution_device
-            assert model.linear2.weight.device == model_device
-        assert model.linear1.weight.device == model_device
-        assert model.batchnorm.weight.device == model_device
-        assert model.linear2.weight.device == model_device
-
-    def test_align_module_offloaded(self):
-        model = ModelForTest()
-        execution_device = torch.device(torch_device)
-        offload_device = torch.device("meta")
-        cpu_offload(model, execution_device=execution_device)
-
-        # test default execution device
-        with align_module(model.batchnorm):
-            assert model.linear1.weight.device == offload_device
-            assert model.batchnorm.weight.device == execution_device
-            assert model.linear2.weight.device == offload_device
-        assert model.linear1.weight.device == offload_device
-        assert model.batchnorm.weight.device == offload_device
-        assert model.linear2.weight.device == offload_device
-
-        # test with explicit execution device
-        with align_module(model.batchnorm, execution_device="cpu"):
-            assert model.linear1.weight.device == offload_device
-            assert model.batchnorm.weight.device == torch.device("cpu")
-            assert model.linear2.weight.device == offload_device
-        assert model.linear1.weight.device == offload_device
-        assert model.batchnorm.weight.device == offload_device
-        assert model.linear2.weight.device == offload_device
